move the license description test file to the components directory, add test cases and remove computed properties from propsData
